Split out Userinterface
