+++ server.py
@@ -13,6 +13,7 @@ class Server:
         self._finishClientId = None
         self._answer = ""
         self._raceSequence = raceSequence
+        self.stop = False
 
     def setCallbackFunctions(self, receivedDataCallback, displayCallback):
         self.receivedDataCallback = receivedDataCallback
@@ -90,69 +91,12 @@ class Server:
             self.display("Wacht op connectie met beide clients")
             time.sleep(2)
 
-    def setNames(self):
-        try:
-            self._raceSequence.racers[0].setName(input("Geef naam van P1...\n"))
-            print("{} Raced als P1".format(self._raceSequence.racers[0].getName()))
-        except:
-            print("Ongeldige naam, instelling niet opgeslagen")
-        try:
-            self._raceSequence.racers[1].setName(input("Geef naam van P2...\n"))
-            print("{} Raced als P2".format(self._raceSequence.racers[1].getName()))
-        except:
-            print("Ongeldige naam, instelling niet opgeslagen")
-
-    def setCountdown(self):
-        try:
-            self._raceSequence.countdown = int(input("Geef aantal seconden...\n"))
-            print("countdown ingesteld op {} seconden".format(self._raceSequence.countdown))
-        except:
-            print("Ongeldig getal, instelling niet opgeslagen ")
-
-    def setOrangeLightAt(self):
-        try:
-            answer = int(input("Geef aantal seconden...\n"))
-            if answer <= self._raceSequence.countdown:
-                self._raceSequence.orangeLightAt = answer
-                print("Oranje licht ingesteld op {} seconden".format(self._raceSequence.orangeLightAt))
-            else:
-                print("Oranje licht sec mag niet hoger zijn dan countdown, instelling niet opgeslagen")
-        except:
-            print("Ongeldig getal, instelling niet opgeslagen")
-
-    def processCommand(self):
-        if self._answer == "start":
-            self.waitForClients()
-            if (not self._answer == 'exit'):
-                self._raceSequence.startRace()
-        elif self._answer == "namen":
-            self.setNames()
-        elif self._answer == "countdown":
-            self.setCountdown()
-        elif self._answer == "oranje":
-            self.setOrangeLightAt()
-        elif self._answer == "help":
-            self.printHelp()
-        elif self._answer == "exit":
-            print("poging tot programma afsluiten, mocht dit niet werken dan programma sluiten met Ctrl+c")
-        else:
-            print("type een geldig commando of type 'help'")
-        self._answer = ""
-
-    def printHelp(self):
-        print("Mogelijke commando's zijn: \n-'start' Om de countdown te beginnen.\n-'namen' Om namen van Racers in te geven.\n-'countdown' Om aantal sec countdown in te stellen.\n-'oranje' Om aantal sec vóór einde countdown in te stellen waarbij het oranje licht aangaat, default is 1.\n-'stop' Om de race af te breken, alleen te gebruiken tijdens de race.\n-'help' Om dit bericht te printen")
-
     def server_program(self):
         start_new_thread(self.keepSteadyConnection, ())
-        start_new_thread(self.waitForAnswer, ())
 
         self.waitForClients()
-        if not self._answer == "exit":
-            self.printHelp()
 
-        while not self._answer == "exit":
+        while not self.stop:
             time.sleep(0.1) # for cpu usage optimization
-            if not self._answer == "":
-                self.processCommand()
 
-        self._raceSequence.stop = True
+
