+++ userInterface.py
@@ -1,0 +1,11 @@
+from raceSequence import RaceSequence
+from server import Server
+
+
+class UserInterface:
+    def __init__(self,server:Server,raceSequence:RaceSequence) -> None:
+        pass
+    def displayText(self,text:str):
+        pass
+    def start(self):
+        pass
